Trying to make go reload work
It doesn't, maybe in the future depending on what happens with the following issue https://github.com/golang/go/issues/11100.
This commit is just to save the workarounds I was trying so that maybe I can come back to it in the future, the problem is that go spawns some threads that don't die after dlclose so segmentation fault happens and I could not find any way to kill these threads. To test run: `cargo run` `curl https://127.0.0.1:7878/api -k -H "Host: jequi.com"` `make reload`